WebServer/requirements.txt
Flask-SocketIO
python-engineio
python-socketio
Flask
Werkzeug
opencv-python==4.5.4.60
numpy
tensorflow
cvlib
sqlalchemy
flask_sqlalchemy  
torch 
torchvision 
torchaudio 
easyocr==1.6.2
Pillow==9.5.0
audiosegment
pytesseract


